Make VM address arithmetic 64-bit-safe (pointer hygiene, stage 3a)
The virtual-memory pager computes page indices by casting pointers to
uint32 and subtracting. On x64 that cast truncates the pointer before
the subtraction, so RequestMemory/bIsMemoryPresent would mislocate (or
Assert on) every page. Widen the pointer casts to a new pointer-width
type so the arithmetic stays correct once the build goes 64-bit.

Adds uintptr/intptr synonyms to UTypes.hpp (the project's types header),
defined via <stdint.h> so they are 32-bit on Win32 and 64-bit on x64.
The page numbers and lengths the arithmetic produces remain uint32 (a
sub-4GB block's offsets fit); an explicit (uint32) narrows them so the
future x64 build is warning-free.

Behaviour-identical on Win32 where uintptr == uint32: same machine code,
verified green (exit 0, no new warnings) and smoke-launched. First slice
of the uint32->uintptr pointer-hygiene sweep; VirtualMem is the
load-bearing one (it is the texture streaming).

+++ jp2_pc/Source/Lib/Sys/VirtualMem.cpp
@@ -274,7 +274,7 @@ void CVirtualMem::SetPagingRegion
 	pu1AllocateBase = (uint8*)(pvBase) + u4AdjustedFileLen;
 
 	// the last address backed by the swap file
-	pvFileEnd = (char*)(((uint32)pvBase) + u4_file_size);
+	pvFileEnd = (char*)(((uintptr)pvBase) + u4_file_size);
 }
